Refactor: 프로필 수정 페이지의 API 분리
